requirements.txt
Flask
flask-cors
python-dotenv
gspread
oauth2client
pandas
pandasql
openai
mcp-server